nightly v1.6.7.0-beta (7/13)
