+++ Assets/Scripts/Camera/CameraController.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using NSMB.Utils;
+using System.Collections.Generic;
 using UnityEngine;
 
 public class CameraController : MonoBehaviour {
@@ -69,7 +70,7 @@ public class CameraController : MonoBehaviour {
         if (playerPos.y - (currentPosition.y - vOrtho) < cameraBottomMax)
             currentPosition.y = playerPos.y + vOrtho - cameraBottomMax;
 
-        float playerHeight = controller.hitboxes[0].size.y * transform.lossyScale.y;
+        float playerHeight = controller.WorldHitboxSize.y;
         float cameraTopMax = Mathf.Min(1.5f + playerHeight, 4f);
 
         //top camera clip
