+++ Assets/Scripts/Entity/Enemy/BobombWalk.cs
@@ -1,6 +1,7 @@
 ﻿using UnityEngine;
 using UnityEngine.Tilemaps;
 using Photon.Pun;
+using NSMB.Utils;
 
 public class BobombWalk : HoldableEntity {
