+++ Assets/Scripts/Entity/Player/DebugControls.cs
@@ -2,6 +2,7 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.Rendering.Universal;
 using Photon.Pun;
+using NSMB.Utils;
 
 public class DebugControls : MonoBehaviour {
