+++ Assets/Scripts/Entity/StarBouncer.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using Photon.Pun;
+using NSMB.Utils;
 
 public class StarBouncer : MonoBehaviourPun {
