+++ Assets/Scripts/Entity/World Elements/FrozenCube.cs
@@ -1,10 +1,9 @@
 using UnityEngine;
 using Photon.Pun;
+using NSMB.Utils;
 
 // maybe a better name for the script
 public class FrozenCube : HoldableEntity {
-
-    private static int GROUND_LAYER_ID = -1;
 
     public float throwSpeed = 10f, shakeSpeed = 1f, shakeAmount = 0.1f;
     public SpriteRenderer spriteRenderer;
@@ -24,9 +23,6 @@ public class FrozenCube : HoldableEntity {
         dead = false;
         holderOffset = Vector2.one;
         body.velocity = Vector2.zero;
-
-        if (GROUND_LAYER_ID == -1)
-            GROUND_LAYER_ID = LayerMask.NameToLayer("Ground");
 
         if (photonView && photonView.InstantiationData != null) {
             int id = (int) photonView.InstantiationData[0];
@@ -127,7 +123,7 @@ public class FrozenCube : HoldableEntity {
             HandleTile();
 
         if (fastSlide && physics.onGround && physics.floorAngle != 0) {
-            RaycastHit2D ray = Physics2D.BoxCast(body.position + Vector2.up * hitbox.size / 2f, hitbox.size, 0, Vector2.down, 0.2f, 1 << GROUND_LAYER_ID);
+            RaycastHit2D ray = Physics2D.BoxCast(body.position + Vector2.up * hitbox.size / 2f, hitbox.size, 0, Vector2.down, 0.2f, Layers.MaskOnlyGround);
             if (ray) {
                 body.position = new Vector2(body.position.x, ray.point.y + Physics2D.defaultContactOffset);
                 if (ray.distance < 0.1f)
@@ -149,7 +145,7 @@ public class FrozenCube : HoldableEntity {
         }
 
         if (throwTimer > 0 && throwTimer - Time.fixedDeltaTime <= 0) {
-            Physics2D.IgnoreCollision(hitbox, previousHolder.hitboxes[0], false);
+            Physics2D.IgnoreCollision(hitbox, previousHolder.MainHitbox, false);
         }
         Utils.TickTimer(ref throwTimer, 0, Time.fixedDeltaTime);
 
@@ -283,7 +279,7 @@ public class FrozenCube : HoldableEntity {
     [PunRPC]
     public override void Pickup(int view) {
         base.Pickup(view);
-        Physics2D.IgnoreCollision(hitbox, holder.hitboxes[0]);
+        Physics2D.IgnoreCollision(hitbox, holder.MainHitbox);
     }
 
     [PunRPC]
