+++ Assets/Scripts/Entity/World Elements/LooseCoin.cs
@@ -1,9 +1,8 @@
 using UnityEngine;
 using Photon.Pun;
+using NSMB.Utils;
 
 public class LooseCoin : MonoBehaviourPun {
-
-    private static int COIN_LAYER = -1, HITSNOTHING_LAYER = -1;
 
     public float despawn = 10;
     private float despawnTimer;
@@ -13,21 +12,16 @@ public class LooseCoin : MonoBehaviourPun {
     private Animator animator;
     private BoxCollider2D hitbox;
 
-    void Start() {
+    public void Start() {
         body = GetComponent<Rigidbody2D>();
         spriteRenderer = GetComponent<SpriteRenderer>();
         hitbox = GetComponent<BoxCollider2D>();
         physics = GetComponent<PhysicsEntity>();
         animator = GetComponent<Animator>();
         body.velocity = Vector2.up * Random.Range(2f, 3f);
-
-        if (COIN_LAYER == -1) {
-            COIN_LAYER = LayerMask.NameToLayer("LooseCoin");
-            HITSNOTHING_LAYER = LayerMask.NameToLayer("HitsNothing");
-        }
     }
 
-    void FixedUpdate() {
+    public void FixedUpdate() {
         if (GameManager.Instance && GameManager.Instance.gameover) {
             body.velocity = Vector2.zero;
             animator.enabled = false;
@@ -36,7 +30,7 @@ public class LooseCoin : MonoBehaviourPun {
         }
 
         bool inWall = Utils.IsAnyTileSolidBetweenWorldBox(body.position + hitbox.offset, hitbox.size * transform.lossyScale * 0.5f);
-        gameObject.layer = inWall ? HITSNOTHING_LAYER : COIN_LAYER;
+        gameObject.layer = inWall ? Layers.LayerHitsNothing : Layers.LayerLooseCoin;
 
         physics.UpdateCollisions();
         if (physics.onGround) {
@@ -46,7 +40,7 @@ public class LooseCoin : MonoBehaviourPun {
         }
 
         spriteRenderer.enabled = !(despawnTimer > despawn-3 && despawnTimer % 0.3f >= 0.15f);
-        
+
         if ((despawnTimer += Time.deltaTime) >= despawn) {
             if (photonView.IsMine)
                 PhotonNetwork.Destroy(photonView);
