+++ Assets/Scripts/GlobalController.cs
@@ -6,6 +6,7 @@ using UnityEngine.SceneManagement;
 using Photon.Pun;
 using Photon.Realtime;
 using ExitGames.Client.Photon;
+using NSMB.Utils;
 
 public class GlobalController : Singleton<GlobalController> {
