+++ Assets/Scripts/Networking/SerializationUtils.cs
@@ -1,3 +1,4 @@
+using NSMB.Utils;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
