+++ Assets/Scripts/Tile/CoinTile.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using Photon.Pun;
+using NSMB.Utils;
 
 [CreateAssetMenu(fileName = "CoinTile", menuName = "ScriptableObjects/Tiles/CoinTile", order = 1)]
 public class CoinTile : BreakableBrickTile {
