+++ Assets/Scripts/Tile/PowerupTile.cs
@@ -1,7 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using Photon.Pun;
+using NSMB.Utils;
 
 [CreateAssetMenu(fileName = "PowerupTile", menuName = "ScriptableObjects/Tiles/PowerupTile", order = 2)]
 public class PowerupTile : BreakableBrickTile {
