+++ Assets/Scripts/UI/Menu/Loading/MarioLoader.cs
@@ -1,22 +1,21 @@
-using System.Collections;
-using System.Collections.Generic;
+using NSMB.Utils;
 using UnityEngine;
 using UnityEngine.UI;
 
 public class MarioLoader : MonoBehaviour {
-    
+
     public int scale = 0, previousScale;
     public float scaleTimer, blinkSpeed = 0.5f;
     private Image image;
     public PlayerData data;
-    void Start() {
+    public void Start() {
         image = GetComponent<Image>();
         data = Utils.GetCharacterData();
     }
 
-    void Update() {
+    public void Update() {
         int scaleDisplay = scale;
-        
+
         if ((scaleTimer += Time.deltaTime) < 0.5f) {
             if (scaleTimer % blinkSpeed < blinkSpeed / 2f)
                 scaleDisplay = previousScale;
