+++ Assets/Scripts/UI/Menu/PlayerListEntry.cs
@@ -3,6 +3,7 @@ using UnityEngine.UI;
 using TMPro;
 
 using Photon.Realtime;
+using NSMB.Utils;
 
 public class PlayerListEntry : MonoBehaviour {
