+++ Assets/Scripts/UI/ScoreboardEntry.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
+using NSMB.Utils;
 
 public class ScoreboardEntry : MonoBehaviour {
